Making model export more object oriented
and some other useless changes
