+++ settings/game.py
@@ -1,11 +1,7 @@
 import os
-import subprocess
-import math
 from pathlib import Path
 import bpy
-import bmesh
-import mathutils
-from .. import common
+
 
 class GameProps(bpy.types.PropertyGroup):
     """Properties for a game"""
@@ -77,72 +73,3 @@ class GameProps(bpy.types.PropertyGroup):
         description="Name of the game you're exporting for",
         default="Game Name",
     )
-
-
-class GameList(bpy.types.UIList):
-    """List of games"""
-    bl_idname = "BASE_UL_GameList"
-
-    def draw_item(self, context, layout, data, item, icon, active_data, active_propname):
-        layout.label(text=item.name)
-
-
-class AddGame(bpy.types.Operator):
-    """Add a game"""
-    bl_idname = "base.add_game"
-    bl_label = "Add Game"
-
-    def execute(self, context):
-        base = context.scene.BASE
-        settings = base.settings
-        settings.games.add()
-        settings.game_index = len(settings.games) - 1
-        return {'FINISHED'}
-
-
-class RemoveGame(bpy.types.Operator):
-    """Remove the selected game from the list"""
-    bl_idname = "base.remove_game"
-    bl_label = "Remove Game"
-
-    @classmethod
-    def poll(cls, context):
-        base = context.scene.BASE
-        settings = base.settings
-        return len(settings.games) > 0
-
-    def execute(self, context):
-        base = context.scene.BASE
-        settings = base.settings
-        settings.games.remove(settings.game_index)
-        settings.game_index = min(
-            max(0, settings.game_index - 1),
-            len(settings.games) - 1
-        )
-        return {'FINISHED'}
-
-
-class MoveGame(bpy.types.Operator):
-    """Move the selected game up or down in the list"""
-    bl_idname = "base.move_game"
-    bl_label = "Move Game"
-
-    direction: bpy.props.EnumProperty(items=(
-        ('UP', "Up", "Move the item up"),
-        ('DOWN', "Down", "Move the item down"),
-    ))
-
-    @classmethod
-    def poll(cls, context):
-        base = context.scene.BASE
-        settings = base.settings
-        return len(settings.games) > 1
-
-    def execute(self, context):
-        base = context.scene.BASE
-        settings = base.settings
-        neighbor = settings.game_index + (-1 if self.direction == 'UP' else 1)
-        settings.games.move(neighbor, settings.game_index)
-        list_length = len(settings.games) - 1
-        settings.game_index = max(0, min(neighbor, list_length))
-        return {'FINISHED'}
