+++ settings/settings.py
@@ -1,10 +1,4 @@
-import os
-import subprocess
-import math
 import bpy
-import bmesh
-import mathutils
-from .. import common
 from . game import GameProps
